+++ @apicize/toolkit/src/controls/main-panel.tsx
@@ -19,12 +19,48 @@ import { CertificateList } from "./navigation/lists/certificate-list";
 import { ProxyList } from "./navigation/lists/proxy-list";
 import { NavigationControl } from "./navigation/navigation";
 import { RequestGroupEditor } from "./editors/request-group-editor";
+import { reaction } from "mobx";
+import { useApicize } from "../contexts/apicize.context";
+import { ToastSeverity, useFeedback } from "../contexts/feedback.context";
+import { useFileOperations } from "../contexts/file-operations.context";
 
 export const MainPanel = observer(() => {
     const workspace = useWorkspace()
+    const settings = useApicize()
+    const fileOps = useFileOperations()
+    const feedback = useFeedback()
 
     const mode = workspace.mode
     const activeSelection = workspace.activeSelection
+
+    let lastPendingChangeCtr = 0
+
+    const checkSave = (pendingChangeCtr: number) => {
+        // Only save if the save ctr hasn't changed in the specified interval
+        setTimeout(async () => {
+            if (lastPendingChangeCtr === pendingChangeCtr) {
+                try {
+                    lastPendingChangeCtr = 0
+                    settings.clearPendingChanges()
+                    await fileOps.saveSettings()
+                } catch (e) {
+                    feedback.toast(`Unable to save Settings - ${e}`, ToastSeverity.Error)
+                }
+            } else {
+                lastPendingChangeCtr = pendingChangeCtr
+                setTimeout(() => {
+                    checkSave(pendingChangeCtr)
+                }, 250)
+            }
+        }, 250)
+    }
+
+    reaction(
+        () => settings.pendingChangeCtr,
+        pendingChangeCtr => {
+            checkSave(pendingChangeCtr)
+        }
+    )
 
     const Pane = (() => {
         switch (mode) {